Fix incorrect error handling within eval.
While processing an eval statement that results in an error,
an internal logic error is produced.

e.g.
    eval('sqrt(-1)')

causes spitbol to exit with the message:

  Failed: internal logic errorsecond second : unexpected ppm branch

While handling a failure in expression evaluation, the routine did not
take into account which stage spitbol was in.

Add a test to see if an eval() is being processed and treat and mark
an error as a failure.