Fix tests for integer divide
The implementation of the minimal instructions dvi and rmi are not
checking the divisor to see if it's zero.  This results in an uncaught
hardware exception, which crashes spitbol itself.

Update the code generated for dvi and rmi to validate the divisor to
ensure it's not zero.

Update the ino, iov, rno and rov instructions to use the 'test' opcode
instead of 'or' instruction.